# chore: use faster `typos` versions